chore: move createCollection to own file
